# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Appointment Booking System",
    "description": "API documentation for the Smart Appointment Booking System",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/api/slots/save": {
      "post": {
        "tags": [
          "Slots"
        ],
        "description": "API route for saving slots",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSlot"
              }
            }
          }
        }
      }
    },
    "/api/slots/get-schedule/{id}": {
      "get": {
        "tags": [
          "Slots"
        ],
        "description": "API route for fetch slots / schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the doctor you want to fetch schedule or availabale slots"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SaveSlot": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "doctor_id": {
              "type": "string",
              "example": "6bddd1a6-81c4-4002-92ca-9e435859e5e5"
            },
            "title": {
              "type": "string",
              "example": "Available Slot"
            },
            "start_time": {
              "type": "string",
              "example": "01:00:00.000"
            },
            "end_time": {
              "type": "string",
              "example": "01:30:00.000"
            },
            "slot_date": {
              "type": "string",
              "example": "2025-06-17"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Enter your token as: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}